+++ packages/producer/src/distributed.ts
@@ -45,6 +45,7 @@ export {
   DEFAULT_CHUNK_SIZE,
   DEFAULT_MAX_PARALLEL_CHUNKS,
   PLAN_DIR_SIZE_LIMIT_BYTES,
+  PLAN_PROJECT_DIR_SKIP_SEGMENTS,
   // Error codes + classes
   FORMAT_NOT_SUPPORTED_IN_DISTRIBUTED,
   FormatNotSupportedInDistributedError,
